Followup compilation fix.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6293 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/gamesys/SaveGame.cpp
@@ -1160,8 +1160,8 @@ void idSaveGame::Write##name_type (const cpp_type value)					\
 		file->Write##name_type(value);										\
 }
 
-DEFINE_READWRITE_PRIMITIVE(int, Int, Long)
-DEFINE_READWRITE_PRIMITIVE(unsigned int, UnsignedInt, Long)
+DEFINE_READWRITE_PRIMITIVE(int, Int, Int)
+DEFINE_READWRITE_PRIMITIVE(unsigned int, UnsignedInt, Int)
 DEFINE_READWRITE_PRIMITIVE(short, Short, Short)
 DEFINE_READWRITE_PRIMITIVE(char, Char, Char)
 DEFINE_READWRITE_PRIMITIVE(unsigned char, UnsignedChar, Char)
